+++ pykrev/__init__.py
@@ -1,3 +1,5 @@
+__version__ = '1.1.2'
+
 from .formula import *
 from .plotting import *
 from .diversity import *