+++ pykrev/formula/__init__.py
@@ -12,3 +12,4 @@ from .read_batch_formularity import read_batch_formularity
 from .calculate_mass import calculate_mass
 from .kendrick_mass_defect import kendrick_mass_defect
 from .filter_spectral_interference import filter_spectral_interference
+from .read_corems import read_corems