+++ pykrev/plotting/kendrick_mass_defect_plot.py
@@ -1,6 +1,6 @@
 from ..formula import kendrick_mass_defect
 from matplotlib import pyplot as plt
-def kendrick_mass_defect_plot(formula_list,mz_list, base = 'CH2', rounding = 'even', **kwargs):
+def kendrick_mass_defect_plot(mz_list, base = 'CH2', rounding = 'even', **kwargs):
     """ 
 	Docstring for function PyKrev.kendrick_mass_defect_plot.py
 	====================
@@ -16,16 +16,17 @@ def kendrick_mass_defect_plot(formula_list,mz_list, base = 'CH2', rounding = 'ev
 	Y: A list of molecular formula. 
     X: A list or numpy array of mz_values of len(Y). 
 	Base: Atom group used to define the Kendrick mass.
+    Rounding: One of 'ceil', 'floor', or 'even', see pk.kendrickMass()
     **kwargs: key word arguments for pyplot.scatter(). 
     Info
 	----------
     Calculation taken from Hughey et al (2001) 
     "Kendrick Mass Defect Spectrum: A Compact Visual Analysis for Ultrahigh-Resolution Broadband Mass Spectra"
     Note: Rounding calclations may lead to artefacts in complex datasets with many peaks. 
-    We recommend experimenting with both even and integer rounding methods when making kmd plots.
+    We recommend experimenting with different rounding methods when making kmd plots.
     """
     
-    kendrickMass, kendrickMassDefect = kendrick_mass_defect(formula_list,mz_list,base=base,rounding=rounding)
+    kendrickMass, kendrickMassDefect = kendrick_mass_defect(mz_list,base=base,rounding=rounding)
     plt.scatter(kendrickMass,kendrickMassDefect, **kwargs)
     plt.xlabel('Kendrick Mass')
     plt.ylabel('Kendrick Mass Defect')